implement containers app, adjust styles to fit grid

+++ js/apps/clockApp.js
@@ -33,7 +33,7 @@ export class ClockApp extends BaseApp {
             }
             if (dateEl) {
                 dateEl.innerText = now.toLocaleDateString([], {
-                    weekday: 'long', month: 'short', day: 'numeric'
+                    weekday: 'short', month: 'short', day: 'numeric'
                 });
             }
         };
